Added address on cin

+++ src/utils/string.js
@@ -1,3 +1,3 @@
 export const arrayToString = (arr) => `${arr.toString()}`
 
-export const coorStrToArray = (str) => str.split(',').map(item => parseInt(item, 10))
+export const coorStrToArray = (str) => str.split(',').map(item => parseFloat(item, 10))
